address review: u32 seeds and validate --seed range
- thread session seed as u32; cast to unsigned int only in plat_seed_rng
- reject negative, signed-prefix, overflow, and out-of-range CLI values

+++ src/platdos.c
@@ -53,7 +53,7 @@ time_t plat_time_now(void)
     return time(NULL);
 }
 
-void plat_seed_rng(unsigned int seed)
+void plat_seed_rng(u32 seed)
 {
-    srand(seed);
+    srand((unsigned int)seed);
 }
